Dispose only Enumerator when Foreach statement is use.
Abstract out Enumerator to be able to dispose of Enumerator rather than whole Shapefile. Before we could only once iterate through collection once because Foreach statement disposing Enumerator on the end of execution.
